refactor: Reorganize components into domain-based hierarchy
Reorganized components directory structure into three categories:
- common/ - Reusable UI components (Button, CheckBoxWithLabel, Dropdown, FloatingButton, Spinner)
- history/ - History domain components (Root, Header, Histories, HistoryItem, SearchBox, etc.)
- settings/ - Settings domain components (OptionsPage, OptionsForm)

Updated all import paths to reflect new structure.
