+++ src/components/history/SearchBox/index.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 
-import { t } from "../../i18n";
+import { t } from "../../../i18n";
 
 interface SearchBoxProps {
   onSearch: (query: string) => void;
